+++ src/main/java/net/silentchaos512/gear/gear/trait/effect/FireproofTraitEffect.java
@@ -1,8 +1,11 @@
 package net.silentchaos512.gear.gear.trait.effect;
 
 import com.mojang.serialization.MapCodec;
+import net.minecraft.core.component.DataComponents;
 import net.minecraft.network.RegistryFriendlyByteBuf;
 import net.minecraft.network.codec.StreamCodec;
+import net.minecraft.util.Unit;
+import net.minecraft.world.item.ItemStack;
 import net.silentchaos512.gear.api.traits.TraitEffect;
 import net.silentchaos512.gear.api.traits.TraitEffectType;
 import net.silentchaos512.gear.setup.gear.TraitEffectTypes;
@@ -29,5 +32,15 @@ public class FireproofTraitEffect extends TraitEffect {
     @Override
     public Collection<String> getExtraWikiLines() {
         return List.of("  - The item cannot be destroyed by fire or lava");
+    }
+
+    @Override
+    public void onRecalculatePre(ItemStack gear, int traitLevel) {
+        gear.remove(DataComponents.FIRE_RESISTANT);
+    }
+
+    @Override
+    public void onRecalculatePost(ItemStack gear, int traitLevel) {
+        gear.set(DataComponents.FIRE_RESISTANT, Unit.INSTANCE);
     }
 }