+++ src/main/java/net/silentchaos512/gear/setup/SgDataComponents.java
@@ -48,16 +48,6 @@ public class SgDataComponents {
                     .persistent(GearPropertiesData.CODEC)
                     .networkSynchronized(GearPropertiesData.STREAM_CODEC)
     );
-    public static final Supplier<DataComponentType<GearPropertiesData>> GEAR_BASE_PROPERTIES = REGISTRAR.registerComponentType(
-            "base_properties",
-            builder -> builder
-                    .networkSynchronized(GearPropertiesData.STREAM_CODEC)
-    );
-    public static final Supplier<DataComponentType<GearPropertyMap>> GEAR_BONUS_PROPERTIES = REGISTRAR.registerComponentType(
-            "bonus_properties",
-            builder -> builder
-                    .networkSynchronized(GearPropertyMap.STREAM_CODEC)
-    );
     @Deprecated // Remove in 1.21.2
     public static final Supplier<DataComponentType<String>> GEAR_MODEL_KEY = REGISTRAR.registerComponentType(
             "model_key",