+++ src/main/java/net/silentchaos512/gear/util/GearData.java
@@ -123,12 +123,10 @@ public final class GearData {
         // All of these are stored for tooltip purposes
         // First, calculate base properties (first pass, creates the traits list and everything)
         var baseProperties = calculateBaseProperties(gear, player, gearType, gearConstructionData);
-        gear.set(SgDataComponents.GEAR_BASE_PROPERTIES, baseProperties);
         // Second, calculate bonus modifiers provided by traits
-        var bonusValues = calculateBonusProperties(gear, player, gearType);
-        gear.set(SgDataComponents.GEAR_BONUS_PROPERTIES, bonusValues);
+        var bonusValues = calculateBonusProperties(gear, player, gearType, baseProperties);
         // Finally, combine the base and bonus modifiers into the final property values
-        var finalProperties = calculateFinalProperties(gear, player, gearType);
+        var finalProperties = calculateFinalProperties(gear, player, gearType, baseProperties, bonusValues);
         gear.set(SgDataComponents.GEAR_PROPERTIES, finalProperties);
 
         printStatsForDebugging(gear, oldProperties, baseProperties, bonusValues, finalProperties);
@@ -209,8 +207,7 @@ public final class GearData {
         return new GearPropertiesData(finalBaseValues);
     }
 
-    private static GearPropertyMap calculateBonusProperties(ItemStack gear, @Nullable Player player, GearType gearType) {
-        var baseProperties = gear.getOrDefault(SgDataComponents.GEAR_BASE_PROPERTIES, GearPropertiesData.EMPTY);
+    private static GearPropertyMap calculateBonusProperties(ItemStack gear, @Nullable Player player, GearType gearType, GearPropertiesData baseProperties) {
         var bonusProperties = new GearPropertyMap();
 
         List<TraitInstance> traits = baseProperties.getOrDefault(GearProperties.TRAITS, TraitListPropertyValue.empty()).value();
@@ -241,10 +238,7 @@ public final class GearData {
         return bonusProperties;
     }
 
-    private static GearPropertiesData calculateFinalProperties(ItemStack gear, @Nullable Player player, GearType gearType) {
-        var baseProperties = gear.getOrDefault(SgDataComponents.GEAR_BASE_PROPERTIES, GearPropertiesData.EMPTY);
-        var bonusProperties = gear.getOrDefault(SgDataComponents.GEAR_BONUS_PROPERTIES, GearPropertyMap.EMPTY);
-
+    private static GearPropertiesData calculateFinalProperties(ItemStack gear, @Nullable Player player, GearType gearType, GearPropertiesData baseProperties, GearPropertyMap bonusProperties) {
         GearPropertyMap combinedMods = new GearPropertyMap();
         Map<GearProperty<?, ?>, GearPropertyValue<?>> finalValues = new LinkedHashMap<>();