+++ src/main/java/net/silentchaos512/gear/data/recipes/ModRecipesProvider.java
@@ -812,6 +812,12 @@ public class ModRecipesProvider extends LibRecipeProvider {
                 .addIngredient(tag(Tags.Items.RODS_BLAZE), 2)
                 .addIngredient(Items.MAGMA_CREAM)
                 .save(this.output);
+
+        CompoundingRecipeBuilder.metalBuilder(CraftingItems.AZURE_ELECTRUM_INGOT, 3)
+                .addIngredient(SgTags.Items.STORAGE_BLOCKS_AZURE_SILVER)
+                .addIngredient(Tags.Items.INGOTS_GOLD, 2)
+                .addIngredient(Tags.Items.ENDER_PEARLS)
+                .save(consumer);
     }
 
     private void registerPressing() {